feat: add update product api

+++ internal/dataaccess/database/migration/000001_init_schema.up.sql
@@ -28,7 +28,7 @@ CREATE TABLE "products" (
   "id" SERIAL PRIMARY KEY,
   "name" VARCHAR(255) NOT NULL,
   "description" TEXT,
-  "price" DECIMAL(10, 2) NOT NULL,
+  "price" float NOT NULL,
   "stock_quantity" INT NOT NULL,
   "status" VARCHAR(50) NOT NULL,
   "image_url" VARCHAR(255),
